# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/sender/webhook/{uid}:
        post:
            tags:
                - Sender
            operationId: Sender_SendWebhook
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendWebhookRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
...
    /v1/webhook/config:
        post:
            tags:
                - Webhook
            operationId: Webhook_CreateWebhook
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.CreateWebhookRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.CreateWebhookReply'
    /v1/webhook/config/{uid}:
        get:
            tags:
                - Webhook
            operationId: Webhook_GetWebhook
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.WebhookItem'
        put:
            tags:
                - Webhook
            operationId: Webhook_UpdateWebhook
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.UpdateWebhookRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.UpdateWebhookReply'
        delete:
            tags:
                - Webhook
            operationId: Webhook_DeleteWebhook
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.DeleteWebhookReply'
    /v1/webhook/config/{uid}/status:
        put:
            tags:
                - Webhook
            operationId: Webhook_UpdateWebhookStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.UpdateWebhookStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.UpdateWebhookStatusReply'
    /v1/webhook/configs:
        get:
            tags:
                - Webhook
            operationId: Webhook_ListWebhook
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: app
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: keyword
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.ListWebhookReply'
    /v1/webhook/configs/select:
        get:
            tags:
                - Webhook
            operationId: Webhook_SelectWebhook
            parameters:
                - name: app
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SelectWebhookReply'
components:
    schemas:
...
        rabbit.api.v1.CreateWebhookReply:
            type: object
            properties: {}
        rabbit.api.v1.CreateWebhookRequest:
            type: object
            properties:
                app:
                    type: integer
                    format: enum
                name:
                    type: string
                url:
                    type: string
                method:
                    type: integer
                    format: enum
                headers:
                    type: object
                    additionalProperties:
                        type: string
                secret:
                    type: string
...
        rabbit.api.v1.DeleteWebhookReply:
            type: object
            properties: {}
...
        rabbit.api.v1.ListWebhookReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/rabbit.api.v1.WebhookItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        rabbit.api.v1.SelectWebhookReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/rabbit.api.v1.WebhookItemSelect'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        rabbit.api.v1.SendReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        rabbit.api.v1.SendWebhookRequest:
            type: object
            properties:
                uid:
                    type: string
                data:
                    type: string
...
        rabbit.api.v1.UpdateWebhookReply:
            type: object
            properties: {}
        rabbit.api.v1.UpdateWebhookRequest:
            type: object
            properties:
                uid:
                    type: string
                app:
                    type: integer
                    format: enum
                name:
                    type: string
                url:
                    type: string
                method:
                    type: integer
                    format: enum
                headers:
                    type: object
                    additionalProperties:
                        type: string
                secret:
                    type: string
        rabbit.api.v1.UpdateWebhookStatusReply:
            type: object
            properties: {}
        rabbit.api.v1.UpdateWebhookStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
        rabbit.api.v1.WebhookItem:
            type: object
            properties:
                uid:
                    type: string
                app:
                    type: integer
                    format: enum
                name:
                    type: string
                url:
                    type: string
                method:
                    type: integer
                    format: enum
                headers:
                    type: object
                    additionalProperties:
                        type: string
                secret:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                status:
                    type: integer
                    format: enum
        rabbit.api.v1.WebhookItemSelect:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
                disabled:
                    type: boolean
                tooltip:
                    type: string
...